Development:Environment - Created an empty database
The database has been created and the structure is generated by
Diesel in Rust from the SQL written to create the migrations.

NOTE: The names of the migration folders do not match the
SQL code for the tables inside because I had to move them. The
tables had to be created in a specific order for the migrations to
make sense and work but I accidentally generated the scaffolding
for them before thinking of that.

No one should need to mess with existing migration SQL but keep
this in mind if you do or need to reference it.